refactor(readr): move HeaderLogo component
