+++ packages/readr/components/layout/layout-with-logo-only.tsx
@@ -3,7 +3,7 @@ import styled from 'styled-components'
 import { SITE_TITLE } from '~/constants/config'
 
 import CustomHead from './custom-head'
-import HeaderLogo from './header-logo'
+import HeaderLogo from './header/header-logo'
 
 const Header = styled.header`
   background-color: #fff;
